Split the mouse support further so the line A owns it. Make work the PS/2 mouse and hook up to EmuTOS and new mouse infrastructure. Atari part is not tested.

+++ include/a2560u.h
@@ -144,7 +144,7 @@ void a2560u_irq_mask_all(uint16_t *save);
 void a2560u_irq_restore(const uint16_t *save);
 void a2560u_irq_enable(uint16_t irq_id);
 void a2560u_irq_disable(uint16_t irq_id);
-void a2560u_irq_acknowledge(uint16_t irq_id);
+void a2560u_irq_acknowledge(uint8_t irq_id);
 void *a2560u_irq_set_handler(uint16_t irq_id, void *handler);
 
 void a2560u_kbd_init(void);
